tidy up committee migrations into own file
